third_party/blink/web_tests/http/tests/security/base-url-data-expected.txt

CONSOLE ERROR: 'data' URLs may not be used as base URLs for a document.
This is a testharness.js-based test.
PASS 'data:' is an invalid base URL.
Harness: the test ran to completion.
